+++ actix-web/src/types/mod.rs
@@ -3,6 +3,7 @@
 mod either;
 mod form;
 mod header;
+mod html;
 mod json;
 mod path;
 mod payload;
@@ -13,6 +14,7 @@ pub use self::{
     either::Either,
     form::{Form, FormConfig, UrlEncoded},
     header::Header,
+    html::Html,
     json::{Json, JsonBody, JsonConfig},
     path::{Path, PathConfig},
     payload::{Payload, PayloadConfig},