+++ awc/examples/client.rs
@@ -1,5 +1,3 @@
-#![allow(clippy::uninlined_format_args)]
-
 use std::error::Error as StdError;
 
 #[tokio::main]