+++ awc/src/responses/response_body.rs
@@ -110,7 +110,7 @@ mod tests {
     use static_assertions::assert_impl_all;
 
     use super::*;
-    use crate::{http::header, test::TestResponse};
+    use crate::test::TestResponse;
 
     assert_impl_all!(ResponseBody<()>: Unpin);